refactor: split universeBuilder and catalogDB into focused modules (#2529)
Splits the ~1900-line server/services/catalogDB.js god-module along its
natural seams into focused sibling modules under catalogDB/, leaving the
original path as a thin re-export barrel so every existing importer keeps
resolving. Pure structural refactor — no behavior change.

- catalogDB/shared.js      row->object mappers, id-minters, bible-payload
                           sanitizer, revision constants, HOMING SQL (leaf)
- catalogDB/scraps.js      scrap CRUD + chunking
- catalogDB/ingredients.js ingredient CRUD, revisions, FTS/vector/hybrid search
- catalogDB/refs.js        source links, external refs, relations, role maps
- catalogDB/media.js       media attachments + integrity surface
- catalogDB/tags.js        tag taxonomy / normalization
- catalogDB/sync.js        peer-sync change feeds + upserts
- catalogDB/facets.js      export slice / sync bundle / stats / facets

Public surface preserved byte-for-byte (65 exports, verified by diff).
Module graph acyclic: shared <- every seam; ingredients <- tags;
facets <- {refs, media}. Full server suite green (only the two pre-existing
nested-worktree install-root tests fail, unrelated to this change).

Closes #2529
