+++ .changelog/NEXT.md
@@ -38,6 +38,7 @@
 
 - **[issue-2528] Collapse the pipeline per-key `createFileWriteQueue` factory into one helper (DRY).** Five pipeline stage stores (`editorialScore.js`, `reverseOutline.js`, `perspectiveRewrite.js`, `manuscriptReview.js`, `continuityBible.js`) each hand-rolled the same `const map = new Map(); if (!map.get(key)) map.set(key, createFileWriteQueue())` get-or-create factory to serialize per-series/per-issue writes against a shared JSON state file. Each now uses a new `createKeyedFileWriteQueue()` in `server/lib/fileWriteQueue.js`, a thin wrapper over the already-existing self-pruning `createKeyCachedQueue` — so the tail Map no longer grows unbounded (idle keys evict on settle) and the five copies collapse to one line apiece. Key normalization (falsy/non-string → shared `'__unknown__'` tail) and observable write ordering are unchanged; all existing stage suites stay green.
 - **[issue-2529] Split the `universeBuilder.js` god-module into focused sibling modules behind a thin barrel.** The ~2000-line `server/services/universeBuilder.js` (50+ exports mixing sanitizers, CRUD, peer-sync, and prompt compilation) was split along its natural seams into `universeBuilder/sanitize.js` (constants + record-shape sanitizers), `universeBuilder/storeFacade.js` (shared synchronous store getter), `universeBuilder/crud.js` (read/list/get, create/update/delete, runs, render-history mutations), `universeBuilder/sync.js` (peer-sync merge + tombstone GC), and `universeBuilder/compile.js` (prompt compilation). The original path is now a 20-line barrel that re-exports the full public surface, so all 94 existing importers keep working unchanged. Pure structural refactor — no behavior change; every existing suite (including the DB-backed tests) stays green. The `catalogDB.js` split is tracked as remaining work under the same issue.
+- **[issue-2529] Split the `catalogDB.js` god-module into focused sibling modules behind a thin barrel.** The ~1900-line `server/services/catalogDB.js` (65 exports mixing row-mappers, scrap/ingredient/tag/media CRUD, refs & relations, peer-sync, and export/facets) was split along its natural seams into `catalogDB/shared.js` (row→object mappers, id-minters, the bible-payload sanitizer, revision constants, and the cross-seam HOMING SQL fragments — the acyclic leaf), `catalogDB/scraps.js` (scrap CRUD + chunking), `catalogDB/ingredients.js` (ingredient CRUD, revisions, FTS/vector/hybrid search), `catalogDB/refs.js` (source links, external refs, relations, role maps), `catalogDB/media.js` (media attachments + integrity surface), `catalogDB/tags.js` (tag taxonomy/normalization), `catalogDB/sync.js` (peer-sync change feeds + upserts), and `catalogDB/facets.js` (export slice / sync bundle / stats / facets). The original path is now a thin barrel that re-exports the identical 65-export public surface, so all existing importers keep working unchanged. Module graph is acyclic (`shared` ← every seam; `ingredients` ← `tags`; `facets` ← {`refs`, `media`}). Pure structural refactor — no behavior change; the full server suite stays green.
 - **[issue-2524] Route per-key write tails through the shared `createKeyCachedQueue` helper (DRY).** Five modules each hand-rolled the same `Map` + `prev.then(fn, fn)` + `.catch`-silence + self-pruning tail to serialize read-modify-write cycles per key: `pipeline/issues.js` (per-series), `loraTraining/db.js` (per-run), `instances.js` (per-peer reciprocal sync), `worktreeManager.js` (per-repo worktree create), and `collectionStore.js` (per-record). Each now delegates to `server/lib/createKeyCachedQueue.js`, and `fileWriteQueue.js`'s id-keyed `createRecordWriteQueue` (already used by the pipeline series/issues and story-builder store facades) is now a thin validator wrapper over the same helper, so all per-key write serialization funnels through one implementation. `collectionStore.js`'s single-tail type-index queue likewise reuses `createFileWriteQueue`. Each site keeps its own key derivation (series-id normalization, macOS realpath resolution) and the fixed reciprocal-sync work body; silence/eviction semantics and observable write ordering are unchanged, and existing write-order tests still pass.
 - **[issue-2523] Extract a shared `createPgFileFacade` backbone for the six PG/file store dispatchers.** The pipeline series/issues, story-builder, universe-builder, catalog-user-types, and writers-room stores each re-implemented the same three pieces: the `isFileBackend()` escape-hatch predicate, the promise-memoized lazy backend selection (so concurrent first calls don't run the migration twice), and the PG bring-up sequence (health-check → `ensureSchema` → one-time file→DB migration → import the leaf-I/O `db.js`). Those now live once in `server/lib/pgFileFacade.js` as `isFileBackend()`, `resolvePgBackend({ requirement, migrate?, loadDb, makePg })`, and `createPgFileFacade({ makeFile, makePg })`; the two inline write-queue copies (`createFileWriteQueue`'s id-keyed sibling `createRecordWriteQueue`) were folded into `server/lib/fileWriteQueue.js`. Each store keeps its own genuinely-different `makeFile`/`makePg` factories and public surface; behavior is preserved on both the PostgreSQL and file/test backends. New unit tests cover the shared helpers.
 - **[issue-2526] Reuse shared helpers instead of local reimplementations (DRY).** Swapped duplicated one-off implementations for their existing `server/lib` counterparts: `sleep` (`youtubeSync.js`, `pipeline/seriesAutopilot.js`) and `pathExists` (`localLlm.js`) now import from `fileUtils.js`; `database.js`'s inline color-only `stripAnsi` uses the shared `ansiStrip.js` (full ANSI/control stripper); `systemHealth.js`'s byte formatter and `huggingFaceCatalog.js`'s (via a thin wrapper that preserves its `null`-on-unknown-size fallback contract) now delegate to `fileUtils.formatBytes`; and the two hand-rolled SSE header blocks now use the canonical `SSE_HEADERS` — `ask.js` keeps its extra `no-transform` directive on top of the shared set, and `mediaJobQueue`'s terminal-event path gains the previously-missing `X-Accel-Buffering: no` so it matches the live-stream path.
